backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
google-generativeai==0.5.4
supabase==2.4.2
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
